+++ src/supervisor.h
@@ -18,10 +18,11 @@
 #ifndef WDOG_SUPERVISOR_H_
 #define WDOG_SUPERVISOR_H_
 
-int supervisor_init   (uev_ctx_t *ctx, int enabled, int realtime, char *script);
-int supervisor_exit   (uev_ctx_t *ctx);
+int supervisor_init         (uev_ctx_t *ctx, int enabled, int realtime, char *script);
+int supervisor_exit         (uev_ctx_t *ctx);
 
-int supervisor_enable (int enable);
+int supervisor_enable       (int enable);
+int supervisor_list_clients (int sd);
 
 #endif /* WDOG_SUPERVISOR_H_ */